UI changes and capability for upload of multiple contracts
Modifications:
• Added capability for handling multiple contract upload in the main file
• Made the UI modifications related to that in home html
• Made several other UI improvements to home html, including:
   • Header name change
   • Added tool capabilities description
   • Added more detailed instructions
   • Defined topic hierarchy using bold, italic and gray text

IMPORTANT TECHNICAL NOTE:
To handle multiple files, I modified the object that defines the names of folders and .pkl objects from the "filename" to the "first_client_name". The first client name will be the first item that the user inputs in the third field in the home page. On the same direction, I modified the place where the uploaded from the "uploaded_files" folder, to a child folder: ./uploaded_files/client_name/pdfs.